% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.package.template-package.R
\docType{package}
\name{r.package.template-package}
\alias{r.package.template-package}
\title{r.package.template: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UCD-SERG/r.package.template}
  \item \url{https://ucd-serg.github.io/r.package.template/}
  \item Report bugs at \url{https://github.com/UCD-SERG/r.package.template/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
